backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
stripe==2.60.0
python-dotenv==1.1.0
jinja2==3.1.2
aiofiles==23.1.0
